Add files via upload
This project is a simple web application that combines user authentication and product management functionalities. Users can log in or sign up to access the product management page where they can view a list of products and add new products. The project is built using HTML, CSS, and JavaScript

User authentication with login and signup forms.
Product management functionality to add and display products.
A clean and responsive user interface.